// ui/src/models/SGMessage.ts
export default class SGMessage {
    Key_20: string;
    Key_22A: string;
    Key_22C: string;
    Key_30T: string;
    Key_52A: string;
    Key_82A: string;
    Key_87A: string;
    Key_77H: string;
    Key_30V: string;
    Key_36: string;
    Key_32B: string;
    Key_57A: string;
    Key_33B: string;
    Key_53A: string;
    Key_56: string;
    Key_57D: string;
    Key_58A: string;
    Key_24D: string;
    status: string;
    clientRef: string;
}